:recycle: Refactoring exports and imports in lib
