+++ src/lib/api/index.ts
@@ -1,0 +1,4 @@
+import EventAPI from './event';
+import PostAPI from './post';
+
+export { EventAPI, PostAPI };
