--- src/lib/index.ts
@@ -1,3 +1,0 @@
-export * from './types/author';
-export * from './types/post';
-export * from './types/event';
